+++ src/main/java/io/github/lucasifce/gamification/domain/model/Professor.java
@@ -13,7 +13,7 @@ public class Professor {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private Integer id;
+	private Long id;
 	
 	@Column(name = "nome", length = 100, nullable = false)
 	private String nome;
@@ -28,7 +28,7 @@ public class Professor {
 	@JoinColumn(name = "usuario_id")
 	private Usuario usuario;
 	
-	public Professor(Integer id, String nome, String email, String telefone, Usuario usuario) {
+	public Professor(Long id, String nome, String email, String telefone, Usuario usuario) {
 		super();
 		this.id = id;
 		this.nome = nome;
@@ -47,11 +47,11 @@ public class Professor {
 
 	public Professor(){}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
