+++ src/main/java/io/github/lucasifce/gamification/domain/repository/ProfesoresRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import io.github.lucasifce.gamification.domain.model.Professor;
 
-public interface ProfesoresRepository extends JpaRepository<Professor, Integer>{
+public interface ProfesoresRepository extends JpaRepository<Professor, Long>{
 
 }
