+++ src/main/java/io/github/lucasifce/gamification/domain/repository/UsuariosRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import io.github.lucasifce.gamification.domain.model.Usuario;
 
-public interface UsuariosRepository extends JpaRepository<Usuario, Integer>{ 
+public interface UsuariosRepository extends JpaRepository<Usuario, Long>{ 
 
 }
